app: fix Python 3 support
